Fix layout of sections
When there is a non-multiple-of-three number of links in a section (e.g., if deleting a few entries from `frontpage.json`) the layout breaks. This is due to the page relying on having a multiple of three number of 4 units wide (`col-md-4`) links.

(The same issue could also arise if the list items were given an additional `col-sm-6` class to be nicer to small screens.)

This fixes the issue by adding having a `row` class for each section.